Removed extra TakeDamage method

+++ LegendOfZelda/Enemies/Hand.cs
@@ -94,7 +94,7 @@ namespace LegendOfZelda.Enemies
             xDir = rand.Next(0, 2); // 0 for x, 1 for y
             yDir = rand.Next(0, 2); // 0 right/down. 1 for left/up
         }
-        public void TakeDamage(float damage)
+        public void TakeDamage(double damage)
         {
             health =- damage;
         }
